day_9/requirements_streamlit.txt
streamlit>=1.28.0
langchain>=0.3.27
langchain-openai>=0.3.30
python-dotenv>=1.0.0
google-generativeai>=0.8.0
